refactor: :recycle: move backup_db_and_json_files procedure to new module
