+++ app/data/scripts/__init__.py
@@ -1,4 +1,5 @@
 # flake8: noqa
+from app.data.scripts.backup_db_and_json_files import backup_db_and_json_files
 from app.data.scripts.bootstrap_unicode_data import bootstrap_unicode_data
 from app.data.scripts.get_xml_unicode_db import download_xml_unicode_database
 from app.data.scripts.parse_xml_unicode_db import parse_xml_unicode_database
